google-adk-maxia/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-adk-maxia"
version = "0.1.0"
description = "MAXIA AI marketplace tools for Google Agent Development Kit (ADK) — swap crypto, rent GPUs, DeFi yields on 15 blockchains"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "MAXIA", email = "ceo@maxiaworld.app"}]
keywords = ["google-adk", "maxia", "ai-marketplace", "crypto", "defi", "gpu", "blockchain", "agent", "a2a", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "google-adk>=1.0.0",
    "httpx>=0.25.0",
]

[project.urls]
Homepage = "https://maxiaworld.app"
Repository = "https://github.com/majorelalexis-stack/maxia"

[tool.setuptools.packages.find]
where = ["src"]
